+++ Chapter8t11/Operator.h
@@ -1,0 +1,6 @@
+#include <iostream>
+
+class Operator {
+    public:
+        void CFour(); // runs Connect Four
+};
